test gestion erreur

+++ push_swap/ft_push_swap.h
@@ -6,7 +6,7 @@
 /*   By: msuter <msuter@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/07 17:56:28 by msuter            #+#    #+#             */
-/*   Updated: 2025/11/25 13:34:44 by msuter           ###   ########.fr       */
+/*   Updated: 2025/11/27 14:08:13 by msuter           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ int		srch_max_in_a(int *tab_a, int size_a);
 int		find_target_index(int *tab_a, int size_a, int value);
 void	reorganise_order_tab_a(int *tab_a, int size_a);
 int		free_all(int *tab_a, int *tab_b);
-int		ft_atol(const char *str);
+long	ft_atol(const char *str);
 int		min_max(long res);
 int		already_sort(int *tab_a, int size_a);
